Move logic out of index files
