+++ packages/core/src/models/index.ts
@@ -23,4 +23,8 @@
   THE SOFTWARE.
 */
 
-export * from './actions';
+export * from './draft4';
+export * from './jsonSchema';
+export * from './jsonSchema4';
+export * from './jsonSchema7';
+export * from './uischema';
