+++ packages/core/src/models/jsonSchema.ts
@@ -22,6 +22,7 @@
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
   THE SOFTWARE.
 */
+
 import { JsonSchema4 } from './jsonSchema4';
 import { JsonSchema7 } from './jsonSchema7';
