+++ packages/core/src/models/uischema.ts
@@ -22,6 +22,7 @@
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
   THE SOFTWARE.
 */
+
 import { JsonSchema } from './jsonSchema';
 
 /**
