+++ packages/core/src/reducers/index.ts
@@ -22,158 +22,13 @@
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
   THE SOFTWARE.
 */
-import { ControlElement, UISchemaElement } from '../models/uischema';
-import {
-  coreReducer,
-  errorAt,
-  errorsAt,
-  extractAjv,
-  extractData,
-  extractSchema,
-  extractUiSchema,
-  JsonFormsCore,
-  subErrorsAt,
-  ValidationMode
-} from './core';
-import {
-  extractDefaultData,
-  JsonFormsDefaultDataRegistryEntry
-} from './default-data';
-import { defaultDataReducer } from '../reducers/default-data';
-import { JsonFormsRendererRegistryEntry, rendererReducer } from './renderers';
-import { JsonFormsState } from '../store';
-import {
-  findMatchingUISchema,
-  JsonFormsUISchemaRegistryEntry,
-  uischemaRegistryReducer,
-  UISchemaTester
-} from './uischemas';
-import {
-  fetchLocale,
-  findLocalizedSchema,
-  findLocalizedUISchema,
-  i18nReducer
-} from './i18n';
 
-import { Generate } from '../generators';
-import { JsonFormsCellRendererRegistryEntry } from './cells';
-import { JsonSchema } from '../models/jsonSchema';
-
-import { cellReducer } from './cells';
-import { configReducer } from './config';
-import get from 'lodash/get';
-import { Ajv } from 'ajv';
-
-export {
-  rendererReducer,
-  cellReducer,
-  coreReducer,
-  i18nReducer,
-  configReducer,
-  UISchemaTester,
-  uischemaRegistryReducer,
-  findMatchingUISchema,
-  JsonFormsUISchemaRegistryEntry
-};
-export { JsonFormsCore, ValidationMode };
-
-export const jsonFormsReducerConfig = {
-  core: coreReducer,
-  renderers: rendererReducer,
-  cells: cellReducer,
-  config: configReducer,
-  uischemas: uischemaRegistryReducer,
-  defaultData: defaultDataReducer,
-  i18n: i18nReducer
-};
-
-export const getData = (state: JsonFormsState) =>
-  extractData(get(state, 'jsonforms.core'));
-export const getSchema = (state: JsonFormsState): JsonSchema =>
-  extractSchema(get(state, 'jsonforms.core'));
-export const getUiSchema = (state: JsonFormsState): UISchemaElement =>
-  extractUiSchema(get(state, 'jsonforms.core'));
-export const getAjv = (
-  state: JsonFormsState
-): Ajv => extractAjv(get(state, 'jsonforms.core'));
-export const getDefaultData = (
-  state: JsonFormsState
-): JsonFormsDefaultDataRegistryEntry[] =>
-  extractDefaultData(get(state, 'jsonforms.defaultData'));
-export const getRenderers = (
-  state: JsonFormsState
-): JsonFormsRendererRegistryEntry[] => get(state, 'jsonforms.renderers');
-export const getCells = (
-  state: JsonFormsState
-): JsonFormsCellRendererRegistryEntry[] => get(state, 'jsonforms.cells');
-export const getUISchemas = (
-  state: JsonFormsState
-): JsonFormsUISchemaRegistryEntry[] => get(state, 'jsonforms.uischemas');
-
-/**
- * Finds a registered UI schema to use, if any.
- * @param schema the JSON schema describing the data to be rendered
- * @param schemaPath the according schema path
- * @param path the instance path
- * @param fallbackLayoutType the type of the layout to use
- * @param control may be checked for embedded inline uischema options
- */
-export const findUISchema = (
-  uischemas: JsonFormsUISchemaRegistryEntry[],
-  schema: JsonSchema,
-  schemaPath: string,
-  path: string,
-  fallbackLayoutType = 'VerticalLayout',
-  control?: ControlElement,
-  rootSchema?: JsonSchema
-): UISchemaElement => {
-  // handle options
-  if (control && control.options && control.options.detail) {
-    if (typeof control.options.detail === 'string') {
-      if (control.options.detail.toUpperCase() === 'GENERATE') {
-        // force generation of uischema
-        return Generate.uiSchema(schema, fallbackLayoutType);
-      }
-    } else if (typeof control.options.detail === 'object') {
-      // check if detail is a valid uischema
-      if (
-        control.options.detail.type &&
-        typeof control.options.detail.type === 'string'
-      ) {
-        return control.options.detail as UISchemaElement;
-      }
-    }
-  }
-  // default
-  const uiSchema = findMatchingUISchema(uischemas)(schema, schemaPath, path);
-  if (uiSchema === undefined) {
-    return Generate.uiSchema(schema, fallbackLayoutType, '#', rootSchema);
-  }
-  return uiSchema;
-};
-
-export const getErrorAt = (instancePath: string, schema: JsonSchema) => (
-  state: JsonFormsState
-) => {
-  return errorAt(instancePath, schema)(state.jsonforms.core);
-};
-
-export { errorsAt };
-
-export const getSubErrorsAt = (instancePath: string, schema: JsonSchema) => (
-  state: JsonFormsState
-) => subErrorsAt(instancePath, schema)(state.jsonforms.core);
-
-export const getConfig = (state: JsonFormsState) => state.jsonforms.config;
-
-export const getLocale = (state: JsonFormsState) =>
-  fetchLocale(get(state, 'jsonforms.i18n'));
-
-export const getLocalizedSchema = (locale: string) => (
-  state: JsonFormsState
-): JsonSchema => findLocalizedSchema(locale)(get(state, 'jsonforms.i18n'));
-
-export const getLocalizedUISchema = (locale: string) => (
-  state: JsonFormsState
-): UISchemaElement =>
-  findLocalizedUISchema(locale)(get(state, 'jsonforms.i18n'));
+export * from './cells';
+export * from './config';
+export * from './core';
+export * from './default-data';
+export * from './i18n';
+export * from './reducers';
+export * from './renderers';
+export * from './selectors';
+export * from './uischemas';
