+++ packages/core/src/reducers/renderers.ts
@@ -22,6 +22,7 @@
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
   THE SOFTWARE.
 */
+
 import { RankedTester } from '../testers';
 import {
   ADD_RENDERER,
@@ -29,7 +30,7 @@ import {
   REMOVE_RENDERER,
   RemoveRendererAction
 } from '../actions';
-import { Reducer } from '../util/type';
+import { Reducer } from '../util';
 
 export interface JsonFormsRendererRegistryEntry {
   tester: RankedTester;
