+++ packages/core/src/util/Formatted.ts
@@ -22,6 +22,7 @@
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
   THE SOFTWARE.
 */
+
 /**
  * Interface for mapping a given type to a formatted string and back.
  */
