+++ packages/core/src/util/index.ts
@@ -22,131 +22,19 @@
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
   THE SOFTWARE.
 */
-import isEmpty from 'lodash/isEmpty';
-import isArray from 'lodash/isArray';
-import includes from 'lodash/includes';
-import find from 'lodash/find';
-import { JsonSchema, Scopable, UISchemaElement } from '../';
-import { resolveData, resolveSchema } from './resolvers';
-import {
-  compose as composePaths,
-  composeWithUi,
-  toDataPath,
-  toDataPathSegments
-} from './path';
-import { isEnabled, isVisible } from './runtime';
-import { Ajv } from 'ajv';
 
-export { createCleanLabel, createLabelDescriptionFrom } from './label';
-
-/**
- * Escape the given string such that it can be used as a class name,
- * i.e. hashes and slashes will be replaced.
- *
- * @param {string} s the string that should be converted to a valid class name
- * @returns {string} the escaped string
- */
-export const convertToValidClassName = (s: string): string =>
-  s.replace('#', 'root').replace(new RegExp('/', 'g'), '_');
-
-export const formatErrorMessage = (errors: string[]) => {
-  if (errors === undefined || errors === null) {
-    return '';
-  }
-
-  return errors.join('\n');
-};
-
-const hasType = (jsonSchema: JsonSchema, expected: string): boolean => {
-  return includes(deriveTypes(jsonSchema), expected);
-};
-
-/**
- * Derives the type of the jsonSchema element
- */
-const deriveTypes = (jsonSchema: JsonSchema): string[] => {
-  if (isEmpty(jsonSchema)) {
-    return [];
-  }
-  if (!isEmpty(jsonSchema.type) && typeof jsonSchema.type === 'string') {
-    return [jsonSchema.type];
-  }
-  if (isArray(jsonSchema.type)) {
-    return jsonSchema.type;
-  }
-  if (
-    !isEmpty(jsonSchema.properties) ||
-    !isEmpty(jsonSchema.additionalProperties)
-  ) {
-    return ['object'];
-  }
-  if (!isEmpty(jsonSchema.items)) {
-    return ['array'];
-  }
-
-  if (!isEmpty(jsonSchema.allOf)) {
-    const allOfType = find(
-      jsonSchema.allOf,
-      (schema: JsonSchema) => deriveTypes(schema).length !== 0
-    );
-
-    if (allOfType) {
-      return deriveTypes(allOfType);
-    }
-  }
-  // ignore all remaining cases
-  return [];
-};
-
-/**
- * Convenience wrapper around resolveData and resolveSchema.
- */
-const Resolve: {
-  schema(
-    schema: JsonSchema,
-    schemaPath: string,
-    rootSchema?: JsonSchema
-  ): JsonSchema;
-  data(data: any, path: string): any;
-} = {
-  schema: resolveSchema,
-  data: resolveData
-};
-export {
-  resolveData,
-  resolveSchema,
-} from './resolvers';
-export { Resolve };
-
-// Paths --
-const fromScopable = (scopable: Scopable) =>
-  toDataPathSegments(scopable.scope).join('.');
-
-const Paths = {
-  compose: composePaths,
-  fromScopable
-};
-export { composePaths, composeWithUi, Paths, toDataPath };
-
-// Runtime --
-const Runtime = {
-  isEnabled(uischema: UISchemaElement, data: any, ajv: Ajv): boolean {
-    return isEnabled(uischema, data,undefined, ajv);
-  },
-  isVisible(uischema: UISchemaElement, data: any, ajv: Ajv): boolean {
-    return isVisible(uischema, data, undefined, ajv);
-  }
-};
-export { isEnabled, isVisible, Runtime, deriveTypes, hasType };
-
-export * from './renderer';
+export * from './array';
 export * from './cell';
-export * from './runtime';
+export * from './combinators';
 export * from './Formatted';
 export * from './ids';
-export * from './validator';
-export * from './combinators';
-export * from './uischema';
-export * from './array';
-export * from './type';
+export * from './label';
+export * from './path';
+export * from './renderer';
+export * from './resolvers';
+export * from './runtime';
 export * from './schema';
+export * from './type';
+export * from './uischema';
+export * from './util';
+export * from './validator';
